Rename packages files.
